fix: split timeline types into client-safe module
Extract TimelineEventData type and formatTimelineEvent into
src/lib/timeline/types.ts (no server imports) so client components
can import without pulling in postgres/server-db modules.
