+++ src/lib/timeline/types.ts
@@ -1,0 +1,74 @@
+/**
+ * Timeline Event Types & Formatting
+ *
+ * Pure types and formatting functions for structured timeline events.
+ * This file has NO server/DB imports and is safe for client components.
+ */
+
+import {
+  STATUS_CONFIG,
+  SEVERITY_CONFIG,
+  PRIORITY_CONFIG,
+  FREQUENCY_CONFIG,
+} from "~/lib/issues/status";
+
+/**
+ * Structured timeline event payload.
+ * Discriminated union on `type` — stored as jsonb in the `event_data` column.
+ */
+export type TimelineEventData =
+  | { type: "assigned"; assigneeName: string }
+  | { type: "unassigned" }
+  | { type: "status_changed"; from: string; to: string }
+  | { type: "severity_changed"; from: string; to: string }
+  | { type: "priority_changed"; from: string; to: string }
+  | { type: "frequency_changed"; from: string; to: string };
+
+/**
+ * Convert a structured timeline event to a human-readable string.
+ * Used by the timeline UI to display system events.
+ */
+export function formatTimelineEvent(event: TimelineEventData): string {
+  switch (event.type) {
+    case "assigned":
+      return `Assigned to ${event.assigneeName}`;
+    case "unassigned":
+      return "Unassigned";
+    case "status_changed":
+      return `Status changed from ${statusLabel(event.from)} to ${statusLabel(event.to)}`;
+    case "severity_changed":
+      return `Severity changed from ${severityLabel(event.from)} to ${severityLabel(event.to)}`;
+    case "priority_changed":
+      return `Priority changed from ${priorityLabel(event.from)} to ${priorityLabel(event.to)}`;
+    case "frequency_changed":
+      return `Frequency changed from ${frequencyLabel(event.from)} to ${frequencyLabel(event.to)}`;
+  }
+}
+
+function statusLabel(value: string): string {
+  const config = (
+    STATUS_CONFIG as Record<string, { label: string } | undefined>
+  )[value];
+  return config?.label ?? value;
+}
+
+function severityLabel(value: string): string {
+  const config = (
+    SEVERITY_CONFIG as Record<string, { label: string } | undefined>
+  )[value];
+  return config?.label ?? value;
+}
+
+function priorityLabel(value: string): string {
+  const config = (
+    PRIORITY_CONFIG as Record<string, { label: string } | undefined>
+  )[value];
+  return config?.label ?? value;
+}
+
+function frequencyLabel(value: string): string {
+  const config = (
+    FREQUENCY_CONFIG as Record<string, { label: string } | undefined>
+  )[value];
+  return config?.label ?? value;
+}
